requirements.txt
streamlit>=1.37.0
google-generativeai>=0.8.0
ezdxf>=1.3.0
matplotlib>=3.8.0
